♻️(frontend) move initiateUpload to its own folder in sideEffects
We mistakenly put `initiateUpload` in directly in `sideEffects`. Move
it to its own folder in `sideEffects/initiateUpload` as we're adding
more side effect functions.
